proposed changes applied
