+++ cr-examples/onnx/src/main/java/oracle/code/onnx/ExplicitOnnxOperators.java
@@ -25,7 +25,9 @@
 
 package oracle.code.onnx;
 
+import java.lang.foreign.ValueLayout;
 import java.util.Optional;
+import java.util.function.Supplier;
 
 class ExplicitOnnxOperators {
 
@@ -74,4 +76,8 @@ class ExplicitOnnxOperators {
     }
 
     // @@@ Constants for value - TENSOR and sparse_value - SPARSE_TENSOR
+
+    public static <T> Tensor<T> If(Tensor<Boolean> cond, Supplier<Tensor<T>> elseBody, Supplier<Tensor<T>> thenBody) {
+        return cond.data().get(ValueLayout.JAVA_BOOLEAN, 0) ? thenBody.get() : elseBody.get();
+    }
 }
